requirements.txt
numpy
trimesh
networkx
scipy
Pillow
PyQt5